Rename files for two services
