+++ projects/fhi-angular-components/src/lib/fhi-time/fhi-weekpicker/services/week-parser-formatter.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { NgbDateParserFormatter, NgbDateStruct } from '@ng-bootstrap/ng-bootstrap';
 
-import { WeekValidationService } from './week-validator.service';
+import { WeekValidationService } from './week-validation.service';
 import { WeekUtilityService } from './week-utility.service';
 
 /**
